Extract StreamExtensions
