+++ test/MsgPack.UnitTest.Mono/MsgPack.UnitTest.Mono.csproj
@@ -1805,6 +1805,9 @@
     <Compile Include="..\MsgPack.UnitTest\SplittingStream.cs">
       <Link>SplittingStream.cs</Link>
     </Compile>
+    <Compile Include="..\MsgPack.UnitTest\StreamExtensions.cs">
+      <Link>StreamExtensions.cs</Link>
+    </Compile>
     <Compile Include="..\MsgPack.UnitTest\StreamPackerTest.cs">
       <Link>StreamPackerTest.cs</Link>
     </Compile>
